refactor(PrettyJSON): Implement a PrettyJSON renderer with baked in "pretty" toggle support for JSON API returns

+++ hexproof/utils/archive.py
@@ -8,24 +8,28 @@ import zipfile
 from pathlib import Path
 
 
-def unpack_tar_gz(path: Path) -> None:
+def unpack_tar_gz(path: Path, remove: bool = True) -> None:
     """Unpacks a `.tar.gz` archive.
 
     Args:
         path: Path to the archive file.
+        remove: Whether to remove the zip after unpacking.
     """
     with tarfile.open(path, 'r:gz') as tar:
         tar.extractall(path=path.parent)
-    os.remove(path)
+    if remove:
+        os.remove(path)
 
 
-def unpack_zip(path: Path) -> None:
+def unpack_zip(path: Path, remove: bool = True) -> None:
     """Unpacks a `.zip` archive.
 
     Args:
         path: Path to the archive file.
+        remove: Whether to remove the zip after unpacking.
     """
     # Unpack the zip file
     with zipfile.ZipFile(path, 'r') as zip_ref:
         zip_ref.extractall(path=path.parent)
-    os.remove(path)
+    if remove:
+        os.remove(path)
